requirements.txt
ansible==3.3.0
ansible-base==2.10.8
cffi==1.14.5
cryptography==3.4.7
Jinja2==2.11.3
libvirt-python==7.2.0
MarkupSafe==1.1.1
packaging==20.9
pycparser==2.20
pyparsing==2.4.7
PyYAML==5.4.1
